refactor: 상품 관리 위젯 분리
