+++ src/refactoring/widget/product-management/index.ts
@@ -1,0 +1,1 @@
+export { ProductManagement } from "./ui";
